requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
openai==1.78.1
python-dotenv==1.0.0
httpx==0.25.0
PyPDF2==3.0.1 